[TEST] Consolidate filename analysis tests and add Config settings unit tests
- Consolidated all filename analysis tests from test_filename_analysis_extended.py into test_filename_analysis.py.
- Deleted the obsolete test_filename_analysis_extended.py file.
- Added direct unit tests for Config._get_setting_bool and Config._get_setting_int as static methods in test_config_types.py.
