+++ tests/test_filename_analysis.py
@@ -4,7 +4,6 @@ from pathlib import Path
 from gptscan import analyze_filename, Config
 
 def test_analyze_filename_rtlo():
-    # RTLO character U+202E
     rtlo_name = "invoice\u202egepj.exe"
     score, msg = analyze_filename(rtlo_name)
     assert score == 1.0
@@ -19,18 +18,15 @@ def test_analyze_filename_double_extension():
     assert score == 0.9
     assert "double extension" in msg
 
-    # Benign double extension should not be flagged (or at least not with 0.9)
     score, msg = analyze_filename("archive.tar.gz")
     assert score == 0.0
 
 def test_analyze_filename_hidden_whitespace():
-    # Deceptive whitespace
     name = "document.pdf          .exe"
     score, msg = analyze_filename(name)
     assert score == 0.9
     assert "whitespace" in msg
 
-    # Normal whitespace
     score, msg = analyze_filename("my script.py")
     assert score == 0.0
 
@@ -44,21 +40,70 @@ def test_analyze_filename_trailing():
     assert "ends with a suspicious space or dot" in msg
 
 def test_analyze_filename_control_chars():
-    # ASCII 7 is Bell
     name = f"evil{chr(7)}.py"
     score, msg = analyze_filename(name)
     assert score == 0.7
     assert "control characters" in msg
 
 def test_is_supported_file_with_suspicious_name():
-    # Ensure Config is initialized
     Config.initialize()
-
-    # Even if .exe is not in extensions_set, it should be supported if name is suspicious
     Config.extensions_set = {".py"}
 
     assert Config.is_supported_file("invoice\u202egepj.exe") is True
     assert Config.is_supported_file("photo.jpg.exe") is True
-
-    # Benign non-supported should still be False
     assert Config.is_supported_file("photo.jpg") is False
+
+def test_analyze_filename_whitespace_double_extension():
+    name = "document.pdf .exe"
+    score, msg = analyze_filename(name)
+    assert score >= 0.9
+    assert "double extension" in msg.lower()
+
+def test_analyze_filename_tab_double_extension():
+    name = "image.png\t.ps1"
+    score, msg = analyze_filename(name)
+    assert score >= 0.9
+    assert "double extension" in msg.lower()
+
+def test_analyze_filename_multiple_dots_double_extension():
+    name = "invoice.pdf....exe"
+    score, msg = analyze_filename(name)
+    assert score >= 0.9
+    assert "double extension" in msg.lower()
+
+def test_analyze_filename_new_deceptive_prefixes():
+    for ext in [".svg", ".html", ".htm"]:
+        name = f"page{ext}.exe"
+        score, msg = analyze_filename(name)
+        assert score >= 0.9
+        assert "double extension" in msg.lower()
+
+def test_analyze_filename_new_exec_extensions():
+    for ext in [".lnk", ".inf"]:
+        name = f"document.pdf{ext}"
+        score, msg = analyze_filename(name)
+        assert score >= 0.9
+        assert "double extension" in msg.lower()
+
+def test_analyze_filename_whitespace_before_single_extension():
+    name = "malware .exe"
+    score, msg = analyze_filename(name)
+    assert score >= 0.5
+    assert "whitespace" in msg.lower() or "space" in msg.lower()
+
+def test_analyze_filename_unicode_invisible_chars():
+    name = "file\u200b.exe"
+    score, msg = analyze_filename(name)
+    assert score >= 0.7
+    assert "invisible" in msg.lower() or "control" in msg.lower()
+
+    name = "invo\u00adice.exe"
+    score, msg = analyze_filename(name)
+    assert score >= 0.7
+    assert "invisible" in msg.lower() or "control" in msg.lower()
+
+def test_analyze_filename_trailing_dots_complex():
+    name = "report.pdf..."
+    score, msg = analyze_filename(name)
+    assert score >= 0.5
+    assert "dot" in msg.lower()
